Added kit_sch and kit_to to runtime app suite

+++ gnd-sys/app/stop_cfs.sh
@@ -2,5 +2,8 @@
 # Assumes only one cFS porcess is running
 echo "Stop cFS Script"
 CFS=`pgrep core`
-echo $CFS
-kill $CFS
+if ! [ -z "$var" ]
+then
+   echo $CFS
+   kill $CFS
+fi
